reorganise graphics tests
